+++ examples/sandbox/src/clientErrors.ts
@@ -1,0 +1,3 @@
+export function isNetworkError(error: unknown): boolean {
+  return error instanceof Error && (error.message.includes('fetch') || error.message.includes('network'));
+}
